Settings: Moved IMEI and SIM status related code from MyDeviceInfoFragment to HardwareInfoFragment
